Fix Kotlin Int to Long argument type mismatch in SalahAlarmReceiver

+++ android/app/src/main/kotlin/com/pfaacodin01/salahsync/SalahAlarmReceiver.kt
@@ -40,7 +40,8 @@ class SalahAlarmReceiver : BroadcastReceiver() {
                 }
 
                 // Schedule Restore Alarm after durationMinutes
-                scheduleRestoreAlarm(context, alarmId, durationMinutes)
+                scheduleRestoreAlarm(context, alarmId, durationMinutes.toLong())
+
             } catch (e: Exception) {
                 e.printStackTrace()
             }
